pyproject.toml
[project]
name = "pythonasia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "hypercorn>=0.18.0",
    "jinja2>=3.1.6",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "qrcode[pil]>=8.2",
]
